Revert "Move detail page related components to "Detail" folder"
This reverts commit 09aabc13ea34a71615e6302b8bdeacb0aabb2702.
