+++ src/components/Annotations/AnnotationButtons.tsx
@@ -1,7 +1,4 @@
-import {
-  projectConfigSelector,
-  useProjectStore,
-} from "../../../stores/project";
+import { projectConfigSelector, useProjectStore } from "../../stores/project";
 import { BrowseScanButtons } from "../Footer/BrowseScanButtons";
 
 export const AnnotationButtons = () => {
