+++ src/components/Annotations/AnnotationFilter.tsx
@@ -1,12 +1,12 @@
 import React from "react";
 import { Button } from "reactions-knaw-huc";
-import { useAnnotationStore } from "../../../stores/annotation";
+import { useAnnotationStore } from "../../stores/annotation";
 import {
   projectConfigSelector,
   translateProjectSelector,
   useProjectStore,
-} from "../../../stores/project";
-import { selectDistinctBodyTypes } from "../../../utils/broccoli";
+} from "../../stores/project";
+import { selectDistinctBodyTypes } from "../../utils/broccoli";
 
 export const AnnotationFilter = () => {
   const [isOpen, setOpen] = React.useState(false);
