+++ src/components/Footer/BrowseScanButtons.tsx
