+++ src/components/Footer/DetailSearchResultsNavigation.tsx
