+++ src/components/Mirador/Mirador.tsx
