+++ src/components/Mirador/defaultMiradorConfig.ts
