+++ src/components/Mirador/zoomAnnoMirador.ts
