+++ src/components/Text/Annotated/LineSegment.tsx
