+++ src/components/Text/Annotated/LineSegmentWithAnnotations.tsx
