+++ src/components/Text/Annotated/LineSegmentsViewer.tsx
