+++ src/components/Text/Annotated/MetadataDetailLabelValues.tsx
