+++ src/components/Text/Annotated/SegmentGroup.tsx
