+++ src/components/Text/Annotated/SegmentedLine.tsx
