+++ src/components/Text/Annotated/details/EntitySummaryProps.ts
