+++ src/components/Text/Annotated/test/resources/dummyLogicalTextAnnotations.ts
