+++ src/components/Text/Annotated/utils/AnnotationSegmenter.spec.ts
