+++ src/components/Text/Annotated/utils/AnnotationSegmenter.ts
