+++ src/components/Text/Annotated/utils/createLineOffsets.spec.ts
