+++ src/components/Text/Annotated/utils/createLineOffsets.ts
