+++ src/components/Text/Annotated/utils/createLineSearchOffsets.spec.ts
