+++ src/components/Text/Annotated/utils/createSearchHighlightOffsets.ts
