+++ src/components/Text/Annotated/utils/detailPath.ts
