+++ src/components/Text/Annotated/utils/groupSegmentsByGroupId.ts
