+++ src/components/Text/Annotated/utils/listAnnotationOffsets.spec.ts
