+++ src/components/Text/Annotated/utils/listOffsetsByChar.ts
