+++ src/components/Text/Annotated/utils/trimMiddle.ts
