+++ src/components/Text/StyledText.tsx
