+++ src/components/Text/TextPanel.tsx
