+++ src/components/Text/TextPanels.tsx
