+++ src/components/Text/ToggleTextPanels.tsx
