+++ src/components/Text/createSearchRegex.tsx
